fix : 관리자 기능 일간 매출, 주간 매출, 월간 매출에 대한 복합 인덱스를 통한 성능 개선 적용

+++ src/main/java/tableOrder/analytics/dto/response/ResponseAnalyticsDto.java
@@ -20,9 +20,9 @@ public class ResponseAnalyticsDto {
         private String period; //기간
         private String menuName;
         private String categoryName;
-        private Integer orderCount;
+        private Long orderCount;
         private Integer freeCount;
-        private Integer amount;
+        private Long amount;
         private Integer rank;
     }
 
@@ -33,9 +33,9 @@ public class ResponseAnalyticsDto {
     public static class SalesDto{
         private String period;
         private Integer orderCount;
-        private Integer totalPrice;
-        private Integer discountPrice;
-        private Integer finalPrice;
+        private Long totalPrice;
+        private Long discountPrice;
+        private Long finalPrice;
     }
